requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.6
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
django-timezone-field==5.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.7.1
idna==3.3
kombu==5.2.4
lxml==4.9.1
numpy==1.23.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pipenv==2022.7.4
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.14.1
xmltodict==0.13.0
